+++ js/utils.js
@@ -6,4 +6,5 @@ const getRandomNumber = (a, b) => {
 };
 
 const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
+
 export {getRandomArrayElement, getRandomNumber};
